Delete folder `src/app/core/expression`
- move `src/app/core/expression/inputservice.js` into `src/app/gui/form/formservice.js`
- move `src/app/core/expression/tabservice.js` into `src/components/GlobalTabs.vue`
